+++ decoder/fortRtree.go
@@ -127,7 +127,7 @@ func addPokestopToTree(pokestop *Pokestop) {
 }
 
 func addGymToTree(gym *Gym) {
-	log.Infof("FortRtree - add gym %d, lat %f lon %f", gym.Id, gym.Lat, gym.Lon)
+	log.Infof("FortRtree - add gym %s, lat %f lon %f", gym.Id, gym.Lat, gym.Lon)
 
 	fortTreeMutex.Lock()
 	fortTree.Insert([2]float64{gym.Lon, gym.Lat}, [2]float64{gym.Lon, gym.Lat}, gym.Id)
