+++ decoder/pokemon.go
@@ -1164,7 +1164,7 @@ func UpdatePokemonRecordWithDiskEncounterProto(ctx context.Context, db db.DbDeta
 	if pokemon == nil || pokemon.isNewRecord() {
 		// No pokemon found
 		diskEncounterCache.Set(encounterId, encounter, ttlcache.DefaultTTL)
-		return fmt.Sprintf("%s Disk encounter without previous GMO - Pokemon stored for later")
+		return fmt.Sprintf("%s Disk encounter without previous GMO - Pokemon stored for later", encounterId)
 	}
 	pokemon.updatePokemonFromDiskEncounterProto(ctx, db, encounter)
 	savePokemonRecord(ctx, db, pokemon)
